+++ color_models/lib/src/models/hsb_color.dart
@@ -141,7 +141,7 @@ class HsbColor extends ColorModel {
   }
 
   @override
-  HsbColor withValues(List<num> values) {
+  HsbColor fromValues(List<num> values) {
     assert(values.length == 3 || values.length == 4);
     assert(values[0] >= 0 && values[0] <= 360);
     assert(values[1] >= 0 && values[1] <= 100);