+++ flutter_color_models/lib/src/models/hsp_color.dart
@@ -1,3 +1,4 @@
+import 'dart:ui' as ui;
 import 'package:flutter/painting.dart' show Color;
 import 'package:color_models/color_models.dart' as cm;
 import '../color_model.dart';
@@ -110,13 +111,24 @@ class HspColor extends cm.HspColor
   }
 
   @override
-  HspColor withValues(List<num> values) {
+  HspColor fromValues(List<num> values) {
     assert(values.length == 3 || values.length == 4);
     assert(values[0] >= 0 && values[0] <= 360);
     assert(values[1] >= 0 && values[1] <= 100);
     assert(values[2] >= 0 && values[2] <= 100);
     if (values.length == 4) assert(values[3] >= 0 && values[3] <= 255);
     return HspColor.fromList(values);
+  }
+
+  @override
+  HspColor withValues(
+      {double? alpha,
+      double? red,
+      double? green,
+      double? blue,
+      ui.ColorSpace? colorSpace}) {
+    Color color = performWithValues(alpha, red, green, blue, colorSpace);
+    return HspColor.fromColor(color);
   }
 
   @override
@@ -205,4 +217,5 @@ class HspColor extends cm.HspColor
 
   @override
   HspColor convert(cm.ColorModel other) => other.toHspColor().cast();
+    
 }