+++ flutter_color_models/lib/src/models/xyz_color.dart
@@ -1,3 +1,4 @@
+import 'dart:ui' as ui;
 import 'package:flutter/painting.dart' show Color;
 import 'package:color_models/color_models.dart' as cm;
 import '../color_model.dart';
@@ -112,13 +113,24 @@ class XyzColor extends cm.XyzColor
   }
 
   @override
-  XyzColor withValues(List<num> values) {
+  XyzColor fromValues(List<num> values) {
     assert(values.length == 3 || values.length == 4);
     assert(values[0] >= 0 && values[0] <= 100);
     assert(values[1] >= 0 && values[1] <= 100);
     assert(values[2] >= 0 && values[2] <= 100);
     if (values.length == 4) assert(values[3] >= 0 && values[3] <= 255);
     return XyzColor.fromList(values);
+  }
+
+  @override
+  XyzColor withValues(
+      {double? alpha,
+      double? red,
+      double? green,
+      double? blue,
+      ui.ColorSpace? colorSpace}) {
+    Color color = performWithValues(alpha, red, green, blue, colorSpace);
+    return XyzColor.fromColor(color);
   }
 
   @override
@@ -204,4 +216,5 @@ class XyzColor extends cm.XyzColor
 
   @override
   XyzColor convert(cm.ColorModel other) => other.toXyzColor().cast();
+    
 }